2nd commit of core logic real estate data scrubber
This script allows me to press a few buttons and automatically scrub fairly large real estate data pulled from CoreLogic RealQuestPro.  This saves myself and employees from needing to manually parse thousands of entries to remove duplicates and international addresses and also format the headers and columns in a fashion to prepare the sheet for mail merge.  This potentially saves hours of work.